+++ game/rect.go
@@ -66,7 +66,7 @@ func (r rectF64) split(rects *[]rectF64) {
 	*rects = append(*rects, r)
 }
 
-func (r rectF64) draw(dst *ebiten.Image, clr color.Color, totalDimension *[2]float64, shadedCorners *[4]uint8, isVertical uint8) {
+func (r rectF64) draw(dst *ebiten.Image, clr color.Color, totalDimension *[2]float64, shadedCorners *[4]float32, isVertical uint8) {
 	op := &ebiten.DrawRectShaderOptions{}
 	op.GeoM.Scale(r.width, r.height)
 	op.GeoM.Translate(r.x, r.y)
